# Add event.namespace support