--- a/lib/index.js
+++ b/lib/index.js
@@ -12,7 +12,12 @@ var observable = function(el) {
    */
   var callbacks = {},
     slice = Array.prototype.slice,
-    onEachEvent = function(e, fn) { e.replace(/\S+/g, fn) }
+    onEachEvent = function(e, fn) { e.replace(/\S+/g, function(name, pos, ns) {
+      name = name.split('.')
+      ns   = name.splice(1).join('.')
+      name = name[0]
+      fn(name, pos, ns)
+    })}
 
   // extend the object adding the observable methods
   Object.defineProperties(el, {
@@ -26,9 +31,10 @@ var observable = function(el) {
       value: function(events, fn) {
         if (typeof fn != 'function')  return el
 
-        onEachEvent(events, function(name, pos) {
+        onEachEvent(events, function(name, pos, ns) {
           (callbacks[name] = callbacks[name] || []).push(fn)
           fn.typed = pos > 0
+          fn.ns = ns
         })
 
         return el
@@ -48,11 +54,11 @@ var observable = function(el) {
       value: function(events, fn) {
         if (events == '*' && !fn) callbacks = {}
         else {
-          onEachEvent(events, function(name) {
-            if (fn) {
+          onEachEvent(events, function(name, pos, ns) {
+            if (fn || ns) {
               var arr = callbacks[name]
               for (var i = 0, cb; cb = arr && arr[i]; ++i) {
-                if (cb == fn) arr.splice(i--, 1)
+                if (cb == fn || ns && cb.ns == ns) arr.splice(i--, 1)
               }
             } else delete callbacks[name]
           })
@@ -100,14 +106,14 @@ var observable = function(el) {
           args[i] = arguments[i + 1] // skip first argument
         }
 
-        onEachEvent(events, function(name) {
+        onEachEvent(events, function(name, pos, ns) {
 
           fns = slice.call(callbacks[name] || [], 0)
 
           for (var i = 0, fn; fn = fns[i]; ++i) {
             if (fn.busy) return
             fn.busy = 1
-            fn.apply(el, fn.typed ? [name].concat(args) : args)
+            if (!ns || fn.ns == ns) fn.apply(el, fn.typed ? [name].concat(args) : args)
             if (fns[i] !== fn) { i-- }
             fn.busy = 0
           }