One of the limitations of (riot) observable is the lack of name spacing. 

This is particularly apparent when attempting to bind multiple tags to a single datastore. For example: 

```javascript
// <tag1>
this.on('mount',function(){
	usersStore.on('changed.tag1', function(items){ 
		this.update() 
	}) 
})
this.on('unmount',function(){
	usersStore.off('changed.tag1')
})

// <tag2>
this.on('mount',function(){
	usersStore.on('changed.tag2', function(items){ 
		this.update() 
	}) 
})
this.on('unmount',function(){
	usersStore.off('changed.tag2')
})
```

With the above example ideally we would want both tags to update when the store emits a "changed" event —currently it requires the store to emit both events: 'changed.tag1' and 'changed.tag2'.

With namespacing both tags will update when a 'changed' event is triggered. Equally important on unmount only the change event related to the tag in question will be removed. All 'changed' events can still be removed by calling .off('changed').

This minor change massively increases the power and flexibility of the emitter with minimal downside. Yes, it's possible that a few current users may see a change in behaviour if they use dots in their events, but I suspect this user count will be low as the dot is typically used to address namespacing (see jQuery as an example: https://api.jquery.com/event.namespace/ ).

This is one of the reasons why I don't use Riots observable. It would be great if it had this feature. 

Thanks.

Another example:

```javascript
var test = {}
observable(test)

test.on('hello',function(a){console.log('hello',a)}) 
test.on('hello.world',function(a){console.log('hello,world',a)}) 

test.trigger('hello','both')
test.trigger('hello.world','one')
test.off('hello.world')
test.trigger('hello','just "hello" ')
```