@GianlucaGuarini 2016-03-25
@Ke- thank you its looks a great feature and it does not add too much code. I would prefer to have a single commit adding this feature and also I would like to test it in a benchmark to check whether the obsevable methods will be slower. Can you squash your changes and make a new pull request please?