@Ke- 2016-03-25
Great; have squashed the pull request into a single commit.

Within the onEachEvent function using:

```javascript
      name = name.split('.')
      ns   = name[1]
      name = name[0]
```

Is marginally faster; however it means events would only support a single dot.