@Ke- 2016-03-25
Changed the behaviour to enforce a single name space.