@Ke- 2016-03-29
Improved performance as per: http://jsperf.com/namespacetest2