@GianlucaGuarini 2016-03-29
could you update the doc as well please?