@GianlucaGuarini 2016-04-02
It seems that this pull request slows down the riot observable. According to the benchmark added to the repo comparing the new observable against the old one, it seems that now it's 2x slower
```
new-observable#simple event x 762,496 ops/sec ±1.00% (90 runs sampled)
new-observable#namespaced event x 759,938 ops/sec ±1.11% (92 runs sampled)
new-observable#multiple events x 190,263 ops/sec ±1.78% (86 runs sampled)
old-observable#simple event x 1,372,204 ops/sec ±0.94% (90 runs sampled)
old-observable#multiple events x 244,157 ops/sec ±2.06% (85 runs sampled)
Fastest is old-observable#simple event
```
Maybe we need to handle this feature only using regex