@Ke- 2016-04-02
Yes, I would expect it to be slower due to the intermediate function. Unfortunately using a more advance regex is slower again —see below. For the absolute best performance we should consider dropping the regexp completely:

```
splice-observable#simple event x 1,377,984 ops/sec ±1.29% (83 runs sampled)
split-observable#namespaced event x 1,835,461 ops/sec ±0.97% (80 runs sampled)
ns-regexp-observable#namespaced event x 1,280,703 ops/sec ±1.21% (84 runs sampled)
for-observable#namespaced event x 3,815,956 ops/sec ±0.98% (81 runs sampled)
for-observable#multiple events x 1,586,047 ops/sec ±1.06% (85 runs sampled)
old-observable#namespaced event x 2,877,195 ops/sec ±0.82% (91 runs sampled)
old-observable#multiple events x 1,539,768 ops/sec ±1.28% (83 runs sampled)
Fastest is for-observable#namespaced event
```

Benchmarks and methods here:

```js
var Benchmark = require('benchmark');

var onEachEventOld = function(e, fn) { e.replace(/\S+/g, fn) }

var onEachEventNSReg = function(e, fn) { e.replace(/(\w+)\.?(\S+)/g, fn) }

var onEachEventSplice = function(e, fn) { e.replace(/\S+/g, function(name, pos, ns) {
  name = name.split('.')
  ns   = name.splice(1).join('.')
  name = name[0]
  fn(name, pos, ns)  
})}

var onEachEventSplit = function(e, fn) { e.replace(/\S+/g, function(name, pos, ns) {
  name = name.split('.')
  ns   = name[1]
  name = name[0]
  fn(name, pos, ns)
})}

var onEachEventFor = function(e, fn) {
  var es = e.split(' '), l = es.length, i = 0, name, ns
  for(; i < l; i++){
    name = es[i].split('.')
    ns   = name[1]
    name = name[0]
    fn(name, i, ns)
  }
}

var suite = new Benchmark.Suite;

suite.add('splice-observable#simple event', function(){
  onEachEventSplice('click', function(){})
})

suite.add('split-observable#namespaced event', function(){
  onEachEventSplit('click.namespace', function(){})
})

suite.add('ns-regexp-observable#namespaced event', function(){
  onEachEventNSReg('click.namespace', function(){})
})

suite.add('for-observable#namespaced event', function(){
  onEachEventFor('click.namespace', function(){})
})

suite.add('for-observable#multiple events', function(){
  onEachEventFor('click click.namespace keyup', function(){})
})

suite.add('old-observable#namespaced event', function(){
  onEachEventOld('click.namespace', function(){})
})

suite.add('old-observable#multiple events', function(){
  onEachEventOld('click click.namespace keyup', function(){})
})

.on('cycle', function(event) {
  console.log(String(event.target));
})
.on('complete', function() {
  console.log('Fastest is ' + this.filter('fastest').map('name'));
})
.run({ 'async': true });
```