@Ke- 2016-04-02
Another alternative which retains compatibility with multiple dots. 

```js
var onEachEventForSlice = function(e, fn) {
  var es = e.split(' '), l = es.length, i = 0, name, indx
  for(; i < l; i++){
  	name = es[i]
    indx = name.indexOf('.')
    fn(name.substring(indx), i, ~indx ? name.slice(indx + 1) : undefined)
  }
}

// for-slice-observable#namespaced event x 3,551,852 ops/sec ±0.89% (90 runs sampled)
```