@GianlucaGuarini 2016-04-04
The last possibility seems to be the best, I will update the code asap and release `riot-observable@2.4.0`