@Ke- 2016-04-05
Any reason why you decided to keep the regexp vs split + for approach? 

Similar question applies splice vs slice — the for + slice approach appears much faster in my testing:

3,551,852 ops/sec (for-slice-observable#namespaced)
    vs 
1,377,984 ops/sec