@GianlucaGuarini 2016-04-06
I didn't update the script yet. Could you make another pull request please? I will be able to work on it again during the weekend