@GianlucaGuarini 2016-07-24
I think this feature will be removed because it causing problems in riot-route https://github.com/riot/route/issues/63. I will remove it from the next riot release making riot-observable faster and more reliable. Thanks for your contribution anyway