@Ke- 2016-07-24
Never mind; just reviewed the performance numbers.

On Mon, Jul 25, 2016 at 9:46 AM Ke Carlton <ke@zeroloop.com> wrote:

> Why not just add a switch / option to disable namespacing?
>
> I don't mind adding that.
>
> On Mon, Jul 25, 2016 at 9:05 AM Gianluca Guarini <notifications@github.com>
> wrote:
>
>> I think this feature will be removed because it causing problems in
>> riot-route riot/route#63 <https://github.com/riot/route/issues/63>. I
>> will remove it from the next riot release making riot-observable faster and
>> more reliable. Thanks for your contribution anyway
>>
>> —
>> You are receiving this because you were mentioned.
>>
>> Reply to this email directly, view it on GitHub
>> <https://github.com/riot/observable/pull/15#issuecomment-234802260>, or mute
>> the thread
>> <https://github.com/notifications/unsubscribe-auth/AFSWde5K6vRJtj8kJtpIQqnuwkVhnFc5ks5qY9N4gaJpZM4H4jss>
>> .
>>
>