@GianlucaGuarini 2016-07-24
@Ke- this was an awesome feature but it was a breaking change that was introduced in a minor release so I will need to remove it also in the next riot@2.6.0 release. I am sure once riot@3.0.0 we can still discuss about this feature ;)